+++ controller/lib/include/clock_domain_descriptor.h
@@ -58,7 +58,7 @@ namespace avdecc_lib
         /**
          * \return The corresponding Clock Sources by index present in the CLOCK DOMAIN.
          */
-        AVDECC_CONTROLLER_LIB32_API virtual uint16_t STDCALL get_clock_source_by_index(uint32_t clk_src_index) = 0;
+        AVDECC_CONTROLLER_LIB32_API virtual uint16_t STDCALL get_clock_source_by_index(size_t clk_src_index) = 0;
 
         /**
          * \return The CLOCK SOURCE index of the requested CLOCK DOMAIN after sending a
